requirements.txt
pandas
openai-whisper
requests